Make publish.sh be directory based, rather than passing files as args
The old usage for the command was ./publish.sh <markdown> <html> <meta>,
but I think the sane assumption that the all the files should preside in
a given directory will make the command easier and more convenient to
use by just passing a directory as the sole argument. This unfortunately
comes with the drawback that we're forced to assume that the <markdown>
and <meta> files are named the same (except for the extension) as the
directory, but I think that's not too unreasonable of a requirement to
ask for.

I also added the ability to provide the header.html and footer.html
directory via a -d/--dir option. This allows users to stow away these
partial html files in a different directory than the publish.sh file
and Markdown.pl presides.